+++ modules/module-json.el
@@ -1,0 +1,7 @@
+;;; module-json.el --- json integration
+
+(use-package json-mode)
+  ; repo: https://github.com/joshwnj/json-mode
+
+(provide 'module-json)
+;;; module-json.el ends here